2025.02.02  - 모터, 온도, dac,   기본 기능 완료  - 명령어 설정 기본 완료

+++ cmdMsg.h
@@ -39,8 +39,8 @@ struct TempProfile {
 
 struct motorProfile {
 
-    uint16_t set_PulseCnt;               // set pulse count
-    uint16_t motorSpeed;
+    uint64_t set_PulseCnt;               // set pulse count
+    uint16_t motorSpeed;                // ¼Óµµ 55 % ~ 150%
 };
 //struct TempProfile TempProfile;
